#57 refactored notifications
